pyproject.toml
[project]
name = "simpson-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "kagglehub>=0.2.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]  
include = ["src*"]